+++ src/contains/circle.rs
@@ -3,8 +3,7 @@ use crate::prelude::*;
 impl ContainsShape for Circle {
     fn contains_circle(&self, circle: &Circle) -> bool {
         let dist = self.center().distance(circle.center()) as isize;
-        let max = (self.radius() as isize) - (circle.radius() as isize);
-        dist < max.abs()
+        dist + circle.radius() as isize <= self.radius() as isize
     }
 
     fn contains_ellipse(&self, ellipse: &Ellipse) -> bool {